pyproject.toml
[project]
name = "voicegrievanceapp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "flask>=3.1.1",
    "lxml>=5.4.0",
    "openai>=1.79.0",
    "pydub>=0.25.1",
    "pymongo>=4.13.0",
    "python-dotenv>=1.1.0",
    "selenium>=4.32.0",
]
